requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
elevenlabs==1.51.0
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
whisper==1.1.10
yt-dlp==2025.1.26
